requirements.txt
django~=4.1.5
djangorestframework~=3.14.0
psycopg2-binary

python-dotenv~=0.21.1